+++ CHANGELOG.md
@@ -16,3 +16,8 @@ This file documents all changes to Argus-frontend. This file is primarily meant 
 
 - Order of events and acknowledgements in feed in detailed incident view. Order is now oldest-first.
 
+
+
+### Added
+- End time to the incident detailed view (for closed incidents only)
+